requirements.txt
python-docx==0.8.11
pandas
pylint
requests
pyyaml
